rename: skill 도메인 dto 리패키징
